Pin numpy version in vcm build
The wheels for vcm are built using an unpinned version of numpy. Numpy
released v1.20.0 which probably breaks ABI compatibility with older
versions. This means that `pip install vcm` now builds `vcm.mappm` with
the new version of numpy, but then install the old version of numpy.

This PR fixes this by pinning the numpy build requirement to be the same
as the one in constraints

Related to #782