Runfv3 cli
The segmented run directory structure is becoming an implied convention for prognostic run outputs, and with time, only more scripts/analyses will start using this format. Currently, the argo workflows are the only supported way to make this outputs, but argo is hard to use for debugging.

This PR concentrates the segmented run logic into the command line script `runfv3.sh`. This tool consists of a small number of subcommands that should be easy to run locally. The argo workflows have been modified to use this tool, which made them much simpler.

Note that runfv3.sh does not provide support for looping, but this is easy to do. Here is an example of how to run two segments:

```
runfv3.sh create outputUrl fv3configFile chunks runFile
runfv3.sh append outputUrl
runfv3.sh append outputUrl
```

Also, the workflows no longer uses a k8s working volume. These are really hard to recreate locally. Now, the state of the workflow is read from the outputUrl.