Specify chunks within UserConfig
Currently chunk sizes for prognostic run output are specified in a yaml file separately from the rest of the run configuration. This makes configuration more challenging, and in particular makes it difficult to implement validation of the chunk sizes. This PR moves the chunk size specification to within the `UserConfig` object which will make implementing a validator easier in a future PR.

Added public API:
- added optional `chunks` Mapping[str, item] item to `DiagnosticsFileConfig`
- new `FortranFileConfig` dataclass, with `name` and `chunks` attributes
- new `fortran_diagnostics` section of `UserConfig` consisting of `FortranFileConfig` specifications

Significant internal changes:
- default chunk sizes for various predefined diagnostic files are now defined in `runtime/diagnostics/default_diagnostics.py` instead of in the `post_process.py` script
- deleted the `nudge-to-obs` argo workflow and associate scripts, since these would be broken after this PR