Increase shm size for k8s run-fv3gfs jobs
Currently C48 runs with 2x2 layout crash with a bus error. Apparently this can be resolved using the `--shm-size` flag for docker run. In the kubernetes context, this is implemented using this suggestion from stack overflow: https://stackoverflow.com/questions/46085748/define-size-for-dev-shm-on-container-engine/46434614#46434614

The fix is tested by running the prognostic run test within the e2e test on 24 cores.

This PR also cleans up a minor issue with the `runfv3` script.

Significant internal changes:
- mount Memory emptydir to /dev/shm
- remove custom cpu/memory request from `create` step, since all this step does is copy a few objects to GCS

Resolves #982 and resolves #782