Use med-sim-pool for workflows that don't use many cores
A lot of workflows that didn't need many cores were running on our `climate-sim-pool`  (26 cores, 30-ish GB ram), and it would be better to run these on the med-sim-pool (8 cores, 32 GB ram). This PR removes the top level tolerations in the `train-diags-prog` and `train-diags-prog-multiple-models` that were making all templates use climate-sim-pool, and changes some of the workflow components that don't need so many cores to use the `med-sim-pool`.