Refactor/test prognostic run steppers (#1033)
The steppers repeated code and were difficult to test. This PR refactors the nudging and machine learning steppers to enable more reuse, pushing the common/messy bits up to the TimeLoop class. This allows a better separation of concerns between the "Steppers" (see the docstrings of `Stepper` and `TimeLoop` for more information).

This more "pure" interface allowed regression testing more of the ML code. The nudging stepper is harder to test since it requires MPI and more data.

I tested each refactor using checksums, and fixed the non-determinism in the scikit-learn based tests.

Also added:
- debugging conveniences
- made `DerivedFV3State` an actual MutableMapping.