Add function to parse FV3 logs
The standard output of the fv3 model contains useful but messy
information. This function loads statistical and time information from
these outputs. It is not intended to parse the more neat statistical
outputs of the python prognostic run.

I added this to the docs, grouping it with the FV3 restart
functionality, which is a similarly idiosyncratic set of outputs
produced by the model.