Build notebook image based on prognostic run (#1060)
To quickly tune the microphysics schemes of FV3, I would like to quickly run FV3 for a few hours, and then plot the global averaged micro-physical quantities. To simplify the scripting it is nice to do this in a jupyter notebook running in the prognostic run image. I added this image `us.gcr.io/vcm-ml/notebook`.

The notebook image is only about 200 MB larger than the standard prognostic run docker image, which is about 4 GB.

This does increase the complexity of the dockerfile somewhat, but I kept the changes in a separate image stage, which can easily be deleted. To counterbalance this, I  cleaned up how dependencies are handled in the Dockerfile. 

Resolves #1044 

Requirement changes:
- Installs jupyter and friends in a separate stage of the prognostic run dockerfile.
  - The licences of the jupyter dependency stack have been already been approved as part of the fv3net environment.