Make vcm.cos_zenith_angle units-aware for DataArray inputs (#1068)
This fixes #1062 by making sure that if latitude and longitude are provided to vcm.cos_zenith_angle with units of radians, they are converted to have units of degrees. If the DataArrays passed to vcm.cos_zenith_angle do not have units, then they are assumed to have units of degrees.