Add zonal avg - pressure plots to prognostic diags  (#1082)
Adds an additional page for figures of 3D variables' zonal and time means. If the verification set has 3D data in the catalog, it will also plot the zonal and time mean biases. In the event that the diag table does not save 3D outputs (in `diags_3d.zarr` and/or `state_after_timestep.zarr`) for the prognostic run, the report will still complete but these figures will be empty. 

Minor change to vcm.interpolate_1d to keep attrs after interpolating.

Added an optional CLI arg for prognostic_run_diags save: --n-jobs to set parallelism for the computation. Defaults to -1 but for short datasets like the integration test this can lead to http errors when multiple processes try to access the data at the same time. I set the integration test to use --n-jobs 4 because this has always worked for me in local tests.