Remove the fv3fit image
This image has been flaky in the tests, and fv3fit is installed in
the prognostic run and fv3net images.

Modify the training argo workflow to use the prognostic run image.

Move the fv3fit documentation construction to the fv3net image.

Run the fv3fit tests via tox in the "unit" tests rule.