synth: add utilities for directories of zarrs and netcdfs (#1122)
I needed to update the synthetic data of the nudged run to develop an hybrid mapper that works on current data, but I realized this is already. Anyway, I noticed an opportunity to remove some repetitive code.

Most of our datasets are directories of zarrs and netcdfs, yet we have no formal way to represent schema of this nature, or generate data. This leads to repetitive path based logic in the test code, and makes it difficult to update datasets. Every test fixture has it's own idiosyncratic way of being updated.

This PR add some python and CLI utilties for working with directories of schema.

I recommend reviewing in order of the commits. the first makes the API, and subsequent ones use it.

Changes:

- some new python functions synth.{read,write,dump,load}_directory_schema.
- new cli for reading data from a directory and saving it locally: synth-read-schema
- convert the nudged data fixture to use this new formalism