fv3fit tests: combine data_path and config fixtures
These are not independent pieces of data, so should not be in separate
fixtures.

This tests assumes that train_config is valid for the data at
data_source_paths, but this is an implementation detail of the fixture
hierarchy.

This test effectively enforces that the train_config and
data_source_paths remain coupled, otherwise it will fail.