add sklearn test case to test_cli.py
it is now easy to add a new test case and delete the old copy/pasted
test.

We have achieved this by decoupling this test as much as possible from
the fixture hiearchy. The fixture likely became more complex because the
ModelTrainingConfig does not cleanly separate model configuration from
data loading configurations, so it was hard to develop independent
fixtures. Unfortunately, because fixtures couple by names rather than
functional calls, it is harder
to do these kinds of refactors and find the actual independent
directions.

Thinking forward, data_source_path and data_source_name are not
independent pieces of data, but we can leave that another time.