combine data loading info into a single fixture
Now every input to `test_training_integration` is an independent piece
of data (except for possibly validation_timesteps...). In other words,
we expect the training to proceed provided that valid values are
provided for data_info and model_info, individually. Model_info and
data_info can be freely combined.

Therefore this test is more likeley to fail when fv3fit.train has bugs,
not when our training configuration code is wrong, or a dataset becomes
out of date.

It's also easy to replace the data_info fixture with a parametrize.

Coincidentally, it appears that _get_model_config has a more cohesive
frontend than `ModelTrainingConfig`. It groups similar pieces of data
together...this suggests a future respecification of the
ModelTrainingConfig.