speed-up load_diagnostics from 2 mins to 5 secs
It took 1 min 40 seconds to load a 2 run dataset.

Most time was spent filling missing variables and dropping variables.

Missing variables mess up the holoviews plots, so I
replaced the existing code with a new object RunDiagnostics that helps
efficiently query the list of xarray datasets for variables when they
are actually used (e.g. in the plots).

The variable dropping could be an issue, but doesn't arise in practice
since variables matching the asked for varfilters (e.g.
"zonal_mean_bias") returns uniquely dimensioned data. If we run into
problems, we can add code to the plotting functions asserting that the
dimensions are the ones expected.