Add validation of time chunk size in runfv3 create (#1133)
Configuring chunk and run length sizes that are valid for multi-segment runs can be challenging. This PR adds a validation of chunk sizes which ensures simulations will be able to run for multiple segments. The validation is done during the runfv3 create step, so only applies for segmented runs, not for runfv3 run-native runs.

Breaking changes:
- all segmented runs (even one-segment runs) must pass this validation.

Significant internal changes:
- add segmented_run sub-package to runtime with validate.py and cli.py modules
- Define a python CLI in runtime.segmented_run.cli that is used within runfv3 create