add vcm.cubedsphere.to_cross
This code was previously in my explore repo, and is pretty handy
for quickly visualizing cubed sphere data.

I used regression testing/hashing since it is hard to verify this works
without the grid data. You'll have to just trust me that this does
correctly unrotate the fields or try it out yourself.