add vcm.cubedsphere.to_cross (#1155)
* add vcm.cubedsphere.to_cross

This code was previously in my explore repo, and is pretty handy
for quickly visualizing cubed sphere data.

I used regression testing/hashing since it is hard to verify this works
without the grid data. You'll have to just trust me that this does
correctly unrotate the fields or try it out yourself.

* add example to fv3viz docs and fix to_cross

after adding an example with real data, I found a visual bug in the
code. I fixed this.

* Update external/fv3viz/examples/plot_cross.py

Co-authored-by: Spencer Clark <spencerkclark@gmail.com>

* Simplify implementation of to_cross

Co-authored-by: Spencer Clark <spencerkclark@gmail.com>