More flexible prognostic run report CLI (#1209)
* enhance performance of diagnostic loading

fsspec.open is very slow since it does many http requests. This code is much
faster.

* Improve ComputeDiagnostics test coverage

Add some tests requiring network access. We can refactor the fixtures to use
local data if these are too slow, or the test data changes in the future.

* Add DiagnosticFolder abstraction

This allows a cleaner separation between detecting rundirs and loading
data from a run directory.

* centralize diagnostic folder detection logic

It should now be possible to implement new versions of
ComputedDiagnosticsList

* add more flexible html report CLI

The prognostic run report only works with computed diagnostics stored in
a single directory. This was inflexible and coupled the data storage
particulars to the generated report.

Add a new CLI which allows generating a list of diagnostics from several
paths. It works like this:

    prognostic_run_diags report-dirs -o report $base/c384-baseline $base/c48-baseline

* Update ComputeDiagnosticsList docstrings

* simplify test_ComputedDiagnosticsList_from_urls

* respond to reviews

* document the new CLI usage

* Update workflows/prognostic_run_diags/docs/usage.rst

Co-authored-by: Oliver Watt-Meyer <oliwm@vulcan.com>

Co-authored-by: Oliver Watt-Meyer <oliwm@vulcan.com>