Isochoric/isobaric transition based on hydrostatic flag (#1223)
This PR ensures that the column integrated heating diagnostics output by prognostic runs and in offline diagnostics respect the isochoric vs isobaric assumption in the dycore. For now, the offline diagnostics hard-code the isochoric assumption (a breaking change). The prognostic runs choose based on the hydrostatic flag (also a breaking change).

* added vcm.calc.thermo.column_integrated_heating_from_isobaric_transition
* added vcm.calc.thermo.column_integrated_heating_from_isochoric_transition
* deleted vcm.calc.thermo.column_integrated_heat
* pass fv_core_nml.hydrostatic flag to ML and nudging stepper initializations so they can compute column integrated heating appropriately