Add hyperparameter config, use generic training interfaces (#1236)
Currently, sklearn and keras training have separate pre-processing logic before their fit routines are called. Their configuration also loosely types the keyword arguments, by accepting kwargs. This opens the possibility for typos in configuration to go unnoticed, and makes it hard to document what options are available. It also puts model-specific concerns into the top-level routines, instead of being able to dispatch those concerns to Estimator classes.

This PR adds the concept of a hyperparameter configuration class for an Estimator, which contains all of the init arguments for that estimator. This is integrated into the TrainingConfig, and is the only part of the configuration that differs between different estimators.

Added public API:
- Added OptimizerConfig and RegularizerConfig which replace the current arguments for optimizer and regularizer on DenseModel

Refactored public API:
- Removed DenseTrainingConfig and SklearnTrainingConfig, added DenseHyperparameters and RandomForestHyperparameters which strongly type the hyperparameter configuration
- Added hyperparameter class to the API for `register_estimator`
- Renamed `get_config_class` to `get_hyperparameter_class`
- Predictor now takes `**kwargs` in its API to satisfy mypy, but raises a TypeError if any are passed

Significant internal changes:
- Modified Estimator classes to take a single "hyperparameters" dataclass argument, instead of multiple keyword arguments
- added RandomForest estimator which wraps the existing SklearnWrapper estimator, and also provides functionality from now-deleted `sklearn/_train.py`
- Removed target_scaler and n_jobs from the SklearnWrapper init, as there was no way to set these from configuration