Add artifacts CLI for searching prognostic reports by run (#1248)
We have many prognostic run reports and it's hard to find the ones relevant to the runs we are interested in. This PR adds a command line tool to search all reports by run url. This is done by looking up the run in an index. The index can be regenerated by passing the -w flag. 

Added public API:
* `artifacts report` sub-command

Refactored public API:
* the existing artifacts CLI was moved to a sub-command at `artifacts ls`