Unify fv3fit training tests using functional interface (#1264)
Fv3fit currently has separate files to test sklearn and keras training. This can make it difficult to add new models, because the model-specific code in these tests means brand new tests must be made. This PR creates tests that work for all sklearn and keras models, and cover the test cases currently covered by the existing tests (reproducibility, dumping, loading) and more (ability to fit the identity function). Tests for new model types are added automatically, using the default hyperparameter values for the new model training type.

Currently fv3fit uses an Estimator class for training. This class is difficult to test because of the existence of mutable state. To avoid testing the dump functionality of Estimator and the behavior of the class when multiple method calls are involved, this PR refactors training to use a "training function", which returns a dumpable/loadable Predictor.

Removing Estimator also aligns better with the current I/O interface, which associates a single class to a string. It is already not possible to save a loadable Predictor to disk unless either it has a dump function (as we do with the sklearn and keras predictor/estimators), or you make use of internal implementation details of `fv3fit.dump` (as is done by EnsemblePredictor).

Refactored public API:
- removed DummyModel, use ConstantOutputPredictor instead
- removed Estimator
- renamed and refactored routines to use a training Callable instead of Estimator, `get_estimator_class` is now `get_training_function` and `register_estimator` is now `register_training_function`
- added training functions for existing Estimator classes
- added `dump` to `Predictor` API
- ConstantOutputPredictor now returns `0.` for outputs whose value has not been set, instead of raising a `KeyError`.

Significant internal changes:
- Refactored and renamed `_RegressorEnsemble` to `_RandomForestEnsemble`
- Replaced keras and sklearn training test files with unified training tests that parameterize better over model type, and use the new configuration classes