Refactor data configuration as needed for use in offline_diags (#1278)
Configuration for data loading is currently very loosely typed, and has some inconsistencies. In particular, the offline diagnostics currently assume that the batches function used has keyword arguments `mapper_function` and `mapper_kwargs`, despite these not being part of the "batch function" API.

This PR creates formal types which will allow refactoring the offline diagnostics to make use of them. The central feature of this PR is the classes `BatchesConfig` and `BatchesFromMapperConfig`. Both of these provide `load_batches` for use in `fv3fit.train`. I will be refactoring offline diags to mainly depend on `load_batches`, but to provide snapshot data if `load_mapper` is available on the BatchesLoader instance used in the configuration. `loaders` is significantly refactored so that a `MapperConfig` class used by `BatchesFromMapperConfig` can be used to formalize the use of a `data_path`.

Added public API:
- Added BatchesLoader abstract base class used by fv3fit, which provides a way to retrieve a sequence of batches
- Added BatchesFromMapperConfig, which formalizes the pattern existing in many batches functions where a mapper is produced and then converted to batches
- Added BatchesConfig which reflects the existing batch function API with no reference to or requirement of a mapper function
- Added `loaders.typing` with Mapper, Batches, and functions which produce them

Refactored public API:
- DataConfig is removed from`fv3fit`, and is replaced by several classes in `loaders`
- consolidated mapper functions to all take in a `data_path` argument, so they can be configured and called consistently by new MapperConfig class
- Removed arguments `rename_vars`, `dim_order`, and `drop_vars` from `open_fine_res_apparent_sources` mapper function
- Refactored call signature of `open_fine_resolution_nudging_hybrid` to avoid Mapping in favor of stronger typing

Significant internal changes:
- Removed `variables` from data configuration classes, instead passing it when batches are retrieved as is required for use of these classes in the offline diagnostics workflow. Added `additional_variables` attribute to `TrainingConfig` to enable this.
- Deleted load_data_sequence routine in fv3fit, instead using `config.load_batches(variables)`
- Replaced many instances of "batch_function" or "batch_kwargs" with "batches_function" and "batches_kwargs", so that "mapper_function" returning "mapper" reflects "batches_function" returning "batches" (not just a single batch).