package prepare-config
Preparing the config previously required knowing the full path to the
prepare_config.py script. This made it hard to use from other working
directories and  tightly couples the argo workflows to the folder
structure of the prognostic run image.

Package this script with setuptools entrypoints so it doens't need so
much TLC to use.