Fix emulation function "get_nc_files" (#1311)
There was a bug i noticed when using nc_dir_to_tf_dataset that stems from a problem with get_nc_files.

If provided with a path to a GCS directory of netCDF paths, the get_nc_files function used the glob method of a filesystem object, which would strip the protocol (gs://) from the list of files. The removal of this information breaks the loading of the remote netCDF downstream because we a) don't provide initiated filesystem objects (gets messy fast) and b) without any protocol info in the path get_remote_nc thinks the file is local. I adjusted get_nc_files to keep the path if it detects a GCS protocol for now.

* Add protocol preservation for globbed paths

* Update Mock name to be more descriptive

* Add missing kwarg to get_nc_files docstring

* Linting

* Remove network mark

* Fix fv3gfs-fortran submodule

* remove pytest import