prog run: raise error for undefined diagnostics
Previously if the user requested saving a diagnostic field not computed
by the model, the model execution would continue. The test suite also
used the no-longer supported `All` variable selection configuration.

Raising an exception will provide the user quicker feedback when they
misconfigure a run or if there is a bug.