Add full-physics emulator training and utilities
Add a machine learning model and training script which supports relative
humidity aware losses and partial fitting (to be used for online
learning in a follow-up PR). The data-structure used by this model
`ThermoBasis` may be useful for thermodynamics-aware model training in
other contexts, it allows e.g. training a model to predict relative
humidity, but with specific humidity as the loss function.

- New ThermoBasis data structure and Models/losses which take it in
- New training etnry