Move input/output variables to hyperparameters (#1342)
Currently input/output variables is defined at the top level in TrainingConfig, and passed to training routines. This results in awkward assertion statements in some training routines that require certain inputs or outputs, or do not allow certain outputs. This PR modifies the handling of variables so that the hyperparameters class is responsible for telling the top level only the information it needs (what variables must be in the dataset for training), allowing each training routine to define how it determines inputs and outputs for its resulting Predictor.

Refactored public API:
- Hyperparameter classes now inherit from an ABC and must have a `variables` attribute or property method.
- input_variables and output_variables are moved from TrainingConfig onto RandomForestHyperparameters and DenseHyperparameters. When passed as top level properties in a configuration yaml, a deprecation warning will be emitted.
- `additional_variables` no longer needs to be specified when training random forests with mass scaling