fv3viz: simpler `plot_cube` API without `mappable_var` (#1373)
The `fv3viz.plot_cube` function required a dataset input that was the output of `fv3viz.mappable_var`, a two-step process that also required correctly setting the kwargs for `mappable_var`. This PR changes `fv3viz` API to use a single call of `plot_cube`, in which now the grid metadata must be specified in order to accomplish the grid and plotting variable dimension alignment that was previously done by `fv3viz.mappable_var`. The default grid metadata in `fv3viz` is from the python wrapper, so for most applications it does not actually need to be provided. It also adds a new public data structure in `vcm.cubedsphere` for storing grid metadata.

Added public API:
- `vcm.cubedsphere.GridMetadata`: a data structure for grid variable name and dimensions

Refactored public API:
- `fv3viz.plot_cube` now takes two required arguments (`ds` and `var_name`), and has a new optional `grid_metadata` arg for when the grid is not the default values, i.e., GFDL FV3 Fortran diagnostic data
- the former `fv3viz.mappable_var` and `fv3viz.plot_cube_axes` are now private 
- removed `fv3viz.plot_diag_var_single_map` which largely duplicated `plot_cube`

Significant internal changes:
- `fv3viz.plot_cube` internally uses `_mappable_var` and the grid metadata to ensure plottable data

- [X] Tests added