fv3fit: sklearn RF serializing metadata as dict (#1404)
The fv3fit `SklearnWrapper` model class dumped a yaml list of metadata instead of a dict, which then required tracking the number and sequence of the items in the list, creating overhead and making breaking changes likely. This PR switches the metadata serialization to a dict, and adds backward compatible loading. Also added a missing base class init call, which sets the input and output variables of `SklearnWrapper` without needing the obsolete properties methods.

Significant internal changes:
- slight changes to way fv3fit RF models are serialized