CircleCI and makefile hooks for image tests
Sometimes, we would like to test an image once it is built rather than including
the tests in the image. For example, docker caching can hide tests failures when
tests access external state. Then the error will appear only when the cache is
broken rather than when the state first changed.

Testing in-dockerfile also makes it hard to bootstrap a test environment. If the
build fails, the image does not exist to debug with.