Add ConvolutionalNetwork and LossConfig classes (#1394)
We are working to add a convolutional model to fv3fit. This PR adds some building blocks to make that happen, particularly new dataclasses ConvolutionalNetwork and LossConfig which can be configured from yaml.

LossConfig duplicates functionality present in slightly different form in two places - the Dense loss, which applies per-feature scaling, and the PrecipitativeModel loss, which scales based on the standard deviation across all features after removing the feature mean. It also adds tests for these different loss options. This new configuration class is assumed to be the loss for a single variable, so the interface for weighting is greatly simplified. Ideally, we should refactor the existing code to use this new configuration class in a later PR.

Added public API:
- Added ConvolutionalNetwork class to configure a sequence of convolutional layers
- Added LossConfig class to configure a loss function for a single variable
- Added numpy_print_precision context manager to `fv3fit.testing`

Significant internal changes:
- Updated unit test code to include "x" and "y" dimensions in test data