prog run: add MergedState object
Python and fortran state are managed differently, which is a bit confusing.
Python state needed to be managed with attributes on the time loop like
`state_updates`, whereas fortran state was put wrapped in a `DerivedFV3State`
object.

Generalize this by adding a MergedState object that stores any python updates
internally.