Emulation layer  and microphsyics model loading/saving (#1410)
This PR is split off from the now closed #1407 to make easier on the review side. It does a few things:

Adds get_config methods to the architecture layers as a step in the direction of having all the layers be serializable without the call-trace.
I had implementations for the fields.py layers, but those ended up being much more complicated because they are a container for other layers. Not essential, but I worked on this because I thought weights and biases needed it for incremental saving. Turns out that's not true.
Added tests that at least confirm the more complicated field i/o layers are saveable in models and reproduce results after loading
Slight adjustments to the layer call signatures to either simplify arguments or set defaults
Added asdict and from_dict methods (and tests) to the MicrophysicsConfig that handles conversion of lists to slices for feature subselection. Slices are not serializable but the sequences are.
Add the ability to turn off/on limiting for the Microphysics models

* Add model/layer serialization adjustments for separate PR

* Linting

* Add a test that checks from_config

* Remove config methods for field layers cuz they don't do anything

* Linting

* Remove positional arg from CombineInputs add get_config test

* undo change to tensorboard

* Undo change to toplevel imports for emulation models

* Linting

* Switch to Oli's slice config style and remove sequence handling

* Actually remove old sequence to slice functions

* Linting