Cleanup prepare config (#1444)
The prepare config script is hard to adapt because there is no clear
specification of the inputs dict. Some of the fields it contains come
from fv3config other from runtime.config, and the handling of
`fortran_diagnostics` and `initial_conditions` overlap in unclear ways.

This formalizes the configuration dictionaries typically used for
prognostic runs using a new class called HighLevelConfig. It combines
fv3confi and runtime configurations and has a higher level interface for
the fortran_diagnostics and initial_conditions.

Added public API:
- `runtime.segmented_run.HighLevelConfig`