Ensure proper Q on report for prescribed/fine-res runs (#1478)
Currently the prognostic report sums the "fv3 physics" and 
"machine learning" tendencies for computing Q1 and Q2. But 
for prescribed/fine-res/emulation runs, the Q1/Q2 should just 
be the ML tendencies and not include the piggy-backed physics. 
This PR allows that by using the 'applied physics' tendency if 
it's available.