relax coupling between diag_table and prognostic run diags (#1495)
The prognostic run diagnostics do not work for arbitrary diag tables and harcodes the names of specific files like `atmos_dt_atmos.zarr` deeply within a stack of helper functions. In practice this tightly couples the diagnostic suite to the configurations for runs and makes it hard to generate reports for different diag tables (e.g. microphysics). For instance, the diagnostics will fail if the input data in `sfc_dt_atmos.zarr`, etc is missing.

This PR relaxes this last assumption and generally cleans-up the code in `load_run_data.py` by removing layers of nearly identical helpers and reformulating the coarsening tools.