Fix microphysics preformance regression (#1521)
Previously, the tendency samples were retrieved via a derived variable in the data transform during loading (far removed from where it was used). The update to the dictionary of tensors brought all the necessary field information into the dict and got rid of a lot of the more complicated name -> list of tensor functions. For incremented fields, the names -> list of tensors included grabbing tendency samples where necessary for proper denormalization. The update didn't have this so denormalizing was to the full field value not the tendency.

Here are the initial loss values:

* Before this commit

18/18 - 1s - loss: 1165458032033792.0000 - air_temperature_output_loss: 2327226880.0000 - cloud_water_mixing_ratio_output_loss: 1.1838 - specific_humidity_output_loss: 3688974.7500 - total_precipitation_loss: 1.1091 - tendency_of_air_temperature_output_mean_squared_error: 0.0000e+00 - tendency_of_specific_humidity_output_mean_squared_error: 0.0000e+00


* After the commit

$ ./projects/microphysics/debug.sh
18/18 - 1s - loss: 90.4563 - air_temperature_output_loss: 1.1115e-04 - cloud_water_mixing_ratio_output_loss: 1.1385 - specific_humidity_output_loss: 6.7398e-05 - total_precipitation_loss: 1.1133 - tendency_of_air_temperature_output_mean_squared_error: 0.0000e+00 - tendency_of_specific_humidity_output_mean_squared_error: 0.0000e+00