Improve prognostic run image build caching (#1524)
* Improve prognostic run image caching

The current dockerfile installs dependencies when it `pip install`s
fv3net code. This means that any change to local fv3net code will force
the docker build to re-download all the dependencies. This makes the
docker builds slow and decreases re-use of existing layers.

Installing external dependencies earlier in the dockerfile improves
caching and decreases the size of the layers used to `pip` install the
local packages.