emulation: run offline/online inside argo workflow
previously the offline and online prognostic runs were in two separate
workflows, and a shell script ensured both of these were called. The diagnostics
workflow required a "baseline" run. This increased the number of parameters
needed by the argo workflow making it harder to use on its own.

In reality for short runs, we want to run both online and offline runs for every
model...and then we can run the prognostic diagnostics comparing these runs,
rather than relying on the external baseline data.