fine_res_budget: add k8s submission manifest
Here is a manifest submitting the job with a pinned version of fv3net.  This
workflow used to be in vcm-workflow-control, but we've started moving workflow
submission manifests into this repo. the process is the same...requires a commit
to change the code and a commit to update the image tag in the k8s yaml. The
only difference is that these commits are to the same repo.