Fix rnn-v1 model reload after save (#1536)
Resolves #1535

The new RNN classes could not be reloaded after saving. This was due to a bug where an architecture layer (RNN) overlapped with a standard name of a keras layer (RNN). During the load of the saved model, keras would try to instantiate the standard layer instead of our custom one.

Notebook showing the error.

Moral of the story, use custom names.

* Failing save/reload test

* Fix bug due to overlapping RNN name

* Linting

* Fix unrecognized arch test

* Change test name to better reflect goal