serialize the transformed_model correctly
Tensorflow uses the first call to a model to infer the input/output signature
for later serialization (see tf.function docs). Unfortunately, the first call to
transformed model included both input and output data, so the serialized model
thinks it requires both inputs and outputs as input!

The functional API is not susceptible to this, because it requires explicitly
constructing the input signature.