Store intermediate ML predictands as diagnostics (#1589)
The transmissivity approach for predicting the downward and net shortwave radiative fluxes depends on first predicting a transmissivity that is neither a state update or a tendency. Before this change, the predicted transmissivity would be incorrectly classified as a tendency in the prognostic run, leading to an error. Now any intermediate variables are simply stored as potential diagnostics.