Prognostic run diags: remove coarsening dimension expansion (#1604)
Currently there's an unintended xarray dimension expansion of the interface grid variables in the prognostic run diags' coarsen-to-C48 step. It doesn't do much damage for C48 inputs but is untenable for C384 inputs. This PR ends this by removing the interface grid variables and coordinates from the prognostic run data on opening (the grid variables are made accessible later in the workflow anyway).