Fix fv3fit.keras.adapters.get_inputs (#1633)
The _convert_dict_output uses the function get_inputs to retrieve the input dictionary for the model, but this was untested. It did not always preserve names for dictionary inputs.

The bug is related to the fact that tf.keras.Model.inputs seems to return a list despite the input format. The previous implementation worked in the case where the layer names already correspond to the dict keys, but failed if input layers just used defaults (e.g., "layer_2"). The fix uses the .input model property instead of .inputs, but I can't seem to find any documentation on these model properties/attributes.

* Failing get_inputs test

* Add passing tests before fix

* Fix failing get_inputs test in keras adapters