Explicitly configure names to be loaded
.backward_names is not sufficient to determine which inputs are internal
or leaf nodes. To see why suppose that a depends on b. Then
composed_transform.backward_names([a, b]) == [a, b].  Clearly there is
no way to use backward_names to infer that a is the only variable that
should be requested from the data.

I explored implementing a Graph class that could keep track of these
arbitrary dependencies, but if felt like overengineering. It is simplest
to just list the required inputs and data variables.